+++ src/documents/item-datamodel.mjs
@@ -1,3 +1,4 @@
+
 /**
  * @extends {foundry.abstract.DataModel}
  */